split up into more components
