+++ src/frontend/src/states/LocalState.tsx
@@ -13,6 +13,7 @@ interface LocalStateProps {
   setHost: (newHost: string, newHostKey: string) => void;
   hostKey: string;
   hostList: HostList;
+  setHostList: (newHostList: HostList) => void;
   language: Locales;
   setLanguage: (newLanguage: Locales) => void;
   // theme
@@ -34,6 +35,7 @@ export const useLocalState = create<LocalStateProps>()(
         set({ host: newHost, hostKey: newHostKey }),
       hostKey: '',
       hostList: {},
+      setHostList: (newHostList) => set({ hostList: newHostList }),
       language: 'en',
       setLanguage: (newLanguage) => set({ language: newLanguage }),
       //theme
